feat: share trip loading and planning logic
